Test ECS reference deployment
Summary:
Use the docker-compose ECS local simulator to test our ECS deployment.
This substitutes the DefaultRunLauncher for the EcsRunLauncher because
the latter cannot run locally and it builds the images from source.

Eventually, we can extend this to run on real ECS as well.

Additionally, this adds a pattern for doing our usual networking
shenanigans but from within PyTest fixtures. I found the approach to
be more flexible and if it's something we like, we can generalize it
and use it across all of our Buildkite docker-compose-in-docker use
cases.

Test Plan: unit, integration

Reviewers: dgibson, alangenfeld, max

Differential Revision: https://dagster.phacility.com/D8575