Refactor instance fixture into an instance factory
This will let us pass overrides for each test. I switched from
`instance_for_test` to `instance_for_test_tempdir` because the former
was making it difficult to properly yield. We were basically running
into this:

```
import os
import tempfile
from contextlib import contextmanager

import pytest

@contextmanager
def yielder():
    with tempfile.TemporaryDirectory() as temp_dir:
        assert os.path.isdir(temp_dir)
        yield temp_dir

@pytest.fixture
def factory():
    def inner():
        with yielder() as yielded:
            return yielded

    yield inner

def test_yielding(factory):
    assert not os.path.isdir(factory())
```

If we return `yielded`, the contextmanager cleans up and the
temporary directory no longer exists.

If we instead yield `yielded`, `factory()` returns a generator and we
need to call `next()` on it for it to be useful.

By passing in our own temporary directory via
`instance_for_test_tempdir`, we can ensure that the temporary directory
still exists even when we return `instance`.

It might be useful to create fixture representations of many of the
functions in test_utils.py but that's out of scope for this change.